fix: Refactor readdir package for updated linting and formatting rules
Introduce new linting and prettier configuration to align with adopted style guides. This update applies the @anolilab configuration to ESLint and Prettier, simplifying the formatting rules and improving code consistency. Additionally, the TypeScript syntax has been refactored in some places and the code optimized to get rid of ESLint warnings. This will make readdir package more maintainable and easier to contribute in