feat: Update dependencies and lockfile version
Updated the versions of several dependencies in various package.json files and the lockfile version in pnpm-lock.yaml. This was to ensure the project uses the latest stable releases of these dependencies for better security, performance, and access to new features. DevDependencies were also updated to their latest versions. Notably, "@types/react" was updated to "18.2.18". In pnpm-lock.yaml, lockfileVersion was downgraded from '6.1' to '6.0', and several package dependencies were also updated.

Please note that this change might impact your project if you are relying on a specific version of a dependency. Verify and validate all relevant dependencies after pulling this commit.

Signed-off-by: prisis <d.bannert@anolilab.de>